ftdisyncbb: Minor changes
