+++ spilpt.ftdisyncbb/compat.h
@@ -1,0 +1,3 @@
+#ifdef COMPAT_USLEEP
+void usleep(int64_t usec);
+#endif
