+++ spilpt.ftdisyncbb/dllmain.c
