--- spilpt.ftdisyncbb/ms_vsnprintf.c
@@ -1,13 +1,0 @@
-/*
- * This fixes linking with pre-compiled libusb-1.0.19-rc1-win: "undefined
- * reference to __ms_vsnprintf". See:
- * http://sourceforge.net/p/mingw-w64/mailman/mingw-w64-public/thread/4F8CA26A.70103@users.sourceforge.net/
- */
-
-#include <stdio.h>
-#include <stdarg.h>
-
-int __ms_vsnprintf(char *str, size_t size, const char *format, va_list ap)
-{
-    return vsnprintf(str, size, format, ap);
-}
