+++ spilpt.ftdisyncbb/spi.c
@@ -14,9 +14,7 @@
 
 #include "spi.h"
 #include "hexdump.h"
-#ifdef WIN_USLEEP_REPLACEMENT
-#include "win_usleep.h"
-#endif
+#include "compat.h"
 
 /* SPI clock frequency. At maximum I got 15KB/s reads at 8 MHz SPI clock. At
  * 12MHz SPI clock it doesn't work. */
