--- spilpt.ftdisyncbb/win_usleep.c
@@ -1,18 +1,0 @@
-/*
- * http://stackoverflow.com/questions/5801813/c-usleep-is-obsolete-workarounds-for-windows-mingw
- */
-#include <windows.h>
-#include <stdint.h>
-
-void usleep(int64_t usec) 
-{ 
-    HANDLE timer; 
-    LARGE_INTEGER ft; 
-
-    ft.QuadPart = -(10*usec); /* Convert to 100 nanosecond interval, negative value indicates relative time */
-
-    timer = CreateWaitableTimer(NULL, TRUE, NULL); 
-    SetWaitableTimer(timer, &ft, 0, NULL, NULL, 0); 
-    WaitForSingleObject(timer, INFINITE); 
-    CloseHandle(timer); 
-}
