--- spilpt.ftdisyncbb/win_usleep.h
@@ -1,1 +1,0 @@
-void usleep(int64_t usec);
